requirements.txt
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.8.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
fastapi==0.115.7
firebase-admin==6.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
msgpack==1.1.0
openai==1.60.1
packaging==24.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
PyPDF2==3.0.1
pytest==8.3.4
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.45.3
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
